+++ CHANGELOG.md
@@ -17,7 +17,8 @@ FEATURES:
 * from/vault: Support Vault >1.9 custom metadata for bundle metadata retrieval. [#68](https://github.com/elastic/harp/pull/68)
 * from/vault: Support legacy bundle metadata format. [#69](https://github.com/elastic/harp/pull/69)
 * template/engine: `jsonEscape` / `jsonUnescape` is added to handle string escaping using JSON character escaping strategy [#70](https://github.com/elastic/harp/pull/70)
-* template/engine: `unquote` is added to unquote a `quote` escaped string [#70](https://github.com/elastic/harp/pull/70)
+* template/engine: `unquote` is added to unquote a `quote` escaped string. [#70](https://github.com/elastic/harp/pull/70)
+* bundle/prefixer: Globally add a prefix to all secret package. [#74](https://github.com/elastic/harp/pull/74)
 
 ## 0.1.24